minor atmos update
Hopefully fixes the atmospherics loop for the ship, so that it does not try to syphon all the air outside of the ship, while filling the ship itself.

Creating weird airflows. so far docking undocking being in space and back has worked and given no problems.

Also fixed a little issue where the blastdoors were not correctly set up for engines and removed the bowman thingy for capn